requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
aiostream==0.6.4
alembic==1.16.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asyncinit==0.2.4
attrs==24.3.0
Automat==24.8.1
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
binance-connector==3.12.0
binance-futures-connector==4.1.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
bokeh @ file:///home/conda/feedstock_root/build_artifacts/bokeh_1747091499565/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229914644/work
cachetools==5.5.2
certifi==2024.12.14
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560551097/work
char==0.1.2
charset-normalizer==3.4.1
clarabel==0.11.1
click==8.1.8
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1736947526808/work
comm==0.2.2
constantly==23.10.4
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1744743161496/work
cryptography==44.0.0
cvxpy==1.6.6
cycler==0.12.1
cytoolz @ file:///Users/runner/miniforge3/conda-bld/cytoolz_1734107258423/work
dask==2025.4.0
databricks-sdk==0.57.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distributed @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_distributed_1751993222/work
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecos==2.0.14
et_xmlfile==2.0.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastexcel==0.14.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
fire==0.7.0
Flask==3.1.1
fonttools==4.56.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
incremental==24.7.2
ipykernel==6.29.5
ipympl==0.9.7
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.10.0
joblib==1.3.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
keyring==25.6.0
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
loguru==0.7.3
lz4 @ file:///Users/runner/miniforge3/conda-bld/lz4_1746561971283/work
Mako==1.3.10
Markdown==3.8.2
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219682943/work
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
ml_dtypes==0.5.1
mlflow==3.1.0
mlflow-skinny==3.1.0
more-itertools==10.6.0
mpire==2.10.2
mpmath==1.3.0
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1749813350845/work
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
narwhals==1.39.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.0
numpy-ext==0.9.9
nvidia-ml-py3==7.352.0
openai==1.95.1
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt_einsum==3.4.0
osqp==1.0.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1715026491486/work
pathspec==0.12.1
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
polars==1.29.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==6.31.1
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1740663164053/work
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pycryptodome==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyportfolioopt==1.5.6
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pyswarms==1.3.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-json-logger==3.3.0
python-okx==0.3.5
python-redis-lock==4.0.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyWavelets==1.8.0
pyxlsb==1.0.10
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454702303/work
pyzmq==27.0.0
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sb3_contrib==2.6.0
scikit-elm==0.21a0
scikit-learn==1.6.1
scikit-opt==0.6.6
scipy==1.15.2
scs==3.2.7.post2
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1738440353519/work
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stable_baselines3==2.6.0
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
stumpy==1.13.0
sympy==1.14.0
ta-lib @ file:///Users/runner/miniforge3/conda-bld/ta-lib_1731944890080/work
tabulate==0.9.0
tblib @ file:///home/conda/feedstock_root/build_artifacts/tblib_1743515515538/work
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tk==0.1.0
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1733736030883/work
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1748003379534/work
tqdm==4.67.1
traitlets==5.14.3
tsfel==0.1.9
tsfresh==0.21.0
Twisted==24.11.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uvicorn==0.34.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
XlsxWriter==3.2.3
xyzservices @ file:///home/conda/feedstock_root/build_artifacts/xyzservices_1745598038694/work
yarl==1.18.3
yfinance==0.2.54
zict @ file:///home/conda/feedstock_root/build_artifacts/zict_1733261551178/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zope.interface==7.2
zstandard==0.23.0
